// src/utils/breakpoint-types.ts
export default [
  'small',
  'medium',
  'mediumAndBelow',
  'mediumAndAbove',
  'large',
  'largeAndBelow',
  'largeAndAbove',
  'xlarge'
]

export type BreakpointEdge =
  | 'small'
  | 'medium'
  | 'mediumAndBelow'
  | 'mediumAndAbove'
  | 'large'
  | 'largeAndBelow'
  | 'largeAndAbove'
  | 'xlarge'

export type Breakpoint = 'small' | 'medium' | 'large' | 'xlarge'